# Optimize direct copy of field name for v1