By default, v1 always specifies EscapeForHTML and EscapeForJS, which prevents the optimization where the pre-quoted name is directly copied to the output buffer.

However, a vast majority of names never have characters that need to be escaped, so pre-compute whether escaping is necessary once within makeStructFields and use the precomputed result. Thus, we speed up marshaling on v1, which has taken a hit in performance when going through the v1in2 emulation.

Performance:

	name                                       old time/op    new time/op    delta
	Testdata/CitmCatalog/Marshal/Concrete      1.03ms ± 0%    0.92ms ± 1%  -10.94%  (p=0.008 n=5+5)
	Testdata/GolangSource/Marshal/Concrete     3.55ms ± 1%    3.20ms ± 2%   -9.88%  (p=0.008 n=5+5)
	Testdata/TwitterStatus/Marshal/Concrete     721µs ± 1%     593µs ± 1%  -17.78%  (p=0.008 n=5+5)